+++ src/main/java/edu/uade/sam/service/DesignService.java
@@ -23,7 +23,7 @@ public interface DesignService {
 	 * @param samples
 	 * @return
 	 */
-	public Design generateDesign(Integer testId, Integer judges, List<String> samples);
+	public Design generateDesign(Long testId, Integer judges, List<String> samples);
 	
 	/**
 	 * Crea el diseño de la prueba con etiquetas random, dando un orden aleatorio a las muestras.
@@ -33,8 +33,8 @@ public interface DesignService {
 	 * @param samples
 	 * @return
 	 */
-	public Design generateDesignRandom(Integer testId, Integer judges, List<String> samples);
+	public Design generateDesignRandom(Long testId, Integer judges, List<String> samples);
 	
-	public Design getTestDesign(Integer testId);
+	public Design getTestDesign(Long testId);
 
 }
